requirements.txt
gunicorn
uvicorn
fastapi==0.110.1
python-dotenv
langchain==0.1.20
langchain-cohere==0.1.3
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.3
langchain-pinecone==0.1.0
langchain-postgres==0.0.3
langchain-text-splitters==0.0.1
pinecone-client
pydantic_settings
pypdf
psycopg
psycopg-binary
passlib
python-jose
bcrypt==3.2.2
pyjwt
pydantic[email]
cohere==5.3.2
httpx[socks]
langsmith
boto3
pytesseract
pdf2image
fastapi-mail==1.4.1
python-multipart
sendgrid
sse-starlette
email-validator
dnspython